+++ src/applications/disability-benefits/996/tests/components/contestedIssues.unit.spec.jsx
@@ -99,33 +99,4 @@ describe('Higher-Level Review 0996 choose contested issues', () => {
       );
     });
   });
-
-  // Office for review section
-  it('should render the same office checkbox', () => {
-    const form = ReactTestUtils.renderIntoDocument(
-      <DefinitionTester
-        definitions={formConfig.defaultDefinitions}
-        schema={schema}
-        data={initialData}
-        uiSchema={uiSchema}
-      />,
-    );
-    const formDOM = getFormDOM(form);
-    expect($('#root_sameOffice', formDOM)).to.not.be.false;
-  });
-
-  // "No" should be selected by default
-  it('should show info alert when checked', () => {
-    const form = ReactTestUtils.renderIntoDocument(
-      <DefinitionTester
-        definitions={formConfig.defaultDefinitions}
-        schema={schema}
-        data={initialData}
-        uiSchema={uiSchema}
-      />,
-    );
-    const formDOM = getFormDOM(form);
-    formDOM.setCheckbox('#root_sameOffice', true);
-    expect($('#root_sameOfficeAlert__title', formDOM)).to.not.be.false;
-  });
 });
